# Add reference count for each lock, free when not used.