@cofyc 2018-07-24
FYI, benchmark on my laptop:

```
$ sysctl -n machdep.cpu.brand_string
Intel(R) Core(TM) i7-4850HQ CPU @ 2.30GHz
$ go test -bench . k8s.io/kubernetes/pkg/util/keymutex
goos: darwin
goarch: amd64
pkg: k8s.io/kubernetes/pkg/util/keymutex
BenchmarkKeyMutexOneKey-8         	 5000000	       307 ns/op
BenchmarkKeyMutexOneKeyWork-8     	 5000000	       384 ns/op
BenchmarkKeyMutexMultiKey-8       	 3000000	       431 ns/op
BenchmarkKeyMutexMultiKeyWork-8   	 3000000	       508 ns/op
PASS
ok  	k8s.io/kubernetes/pkg/util/keymutex	8.942s
```