@msau42 2018-07-24
Awesome! Is it possible to compare the benchmark without the change?