@cofyc 2018-07-24
Without the change:

```
$ sysctl -n machdep.cpu.brand_string
Intel(R) Core(TM) i7-4850HQ CPU @ 2.30GHz
$ go test -bench . k8s.io/kubernetes/pkg/util/keymutex
goos: darwin
goarch: amd64
pkg: k8s.io/kubernetes/pkg/util/keymutex
BenchmarkKeyMutexOneKey-8         	10000000	       213 ns/op
BenchmarkKeyMutexOneKeyWork-8     	 5000000	       279 ns/op
BenchmarkKeyMutexMultiKey-8       	 5000000	       282 ns/op
BenchmarkKeyMutexMultiKeyWork-8   	 5000000	       347 ns/op
PASS
ok  	k8s.io/kubernetes/pkg/util/keymutex	8.845s
```