@cofyc 2018-07-24
Oh, I found most expensive code is `glog.V(5)` statements, even if not logs printed.

In previous tests, I removed `glog` statements in `keymutex`, after adding `glog` statements back:

Without this patch:

```
$ go test -bench . k8s.io/kubernetes/pkg/util/keymutex
goos: darwin
goarch: amd64
pkg: k8s.io/kubernetes/pkg/util/keymutex
BenchmarkKeyMutexOneKey-8         	 2000000	       584 ns/op
BenchmarkKeyMutexOneKeyWork-8     	 2000000	       710 ns/op
BenchmarkKeyMutexMultiKey-8       	 3000000	       585 ns/op
BenchmarkKeyMutexMultiKeyWork-8   	 2000000	       655 ns/op
PASS
ok  	k8s.io/kubernetes/pkg/util/keymutex	9.336s
```

With this patch:

```
$ go test -bench . k8s.io/kubernetes/pkg/util/keymutex
goos: darwin
goarch: amd64
pkg: k8s.io/kubernetes/pkg/util/keymutex
BenchmarkKeyMutexOneKey-8         	 2000000	       622 ns/op
BenchmarkKeyMutexOneKeyWork-8     	 2000000	       770 ns/op
BenchmarkKeyMutexMultiKey-8       	 2000000	       741 ns/op
BenchmarkKeyMutexMultiKeyWork-8   	 2000000	       843 ns/op
PASS
ok  	k8s.io/kubernetes/pkg/util/keymutex	10.010s
```