@msau42 2018-07-25
FYI @bsalamat saw a similar glog issue related to the vmodule setting during performance benchmarks.  Bobby, were you able to disable vmodule in your tests?