@thockin 2018-07-27
Could they all just use a single big lock instead?

On Fri, Jul 27, 2018 at 3:31 PM Michelle Au <notifications@github.com>
wrote:

> *@msau42* commented on this pull request.
> ------------------------------
>
> In pkg/util/keymutex/keymutex.go
> <https://github.com/kubernetes/kubernetes/pull/66442#discussion_r205914449>
> :
>
> >  		return fmt.Errorf("id %q not found", id)
>  	}
>  	glog.V(5).Infof("UnlockKey(...) for id. Mutex found, trying to unlock it. %q\r\n", id)
>
>  	mutex.Unlock()
>  	glog.V(5).Infof("UnlockKey(...) for id %q completed.\r\n", id)
> +	km.RUnlock()
> +
> +	km.Lock()
>
> The local plugin uses it to serialize the fsGroup conflict check, if
> multiple pods share the same volume.
>
> Because the local plugin doesn't implement the Attacher interface,
> MountVolume can be called in parallel
> <https://github.com/kubernetes/kubernetes/blob/master/pkg/volume/util/operationexecutor/operation_executor.go#L744>.
> However, there is work being done to introduce a Local Attacher
> <https://github.com/kubernetes/community/pull/1683> volume interface, and
> move local volumes to implement that. When that happens, then MountVolume
> will be serialized per volume, and then we should not need this lock
> anymore.
>
> I'm not sure about the other plugins, but I think they don't implement
> Attacher either, and use the lock to serialize mount refcounting.
>
> —
> You are receiving this because your review was requested.
> Reply to this email directly, view it on GitHub
> <https://github.com/kubernetes/kubernetes/pull/66442#discussion_r205914449>,
> or mute the thread
> <https://github.com/notifications/unsubscribe-auth/AFVgVKqRQiRcPDuobvvTiqxHb2SLuCg5ks5uK5S-gaJpZM4VXzhH>
> .
>