@thockin 2018-07-27
Or perhaps a simpler abstraction?

On Fri, Jul 27, 2018 at 4:25 PM Michelle Au <notifications@github.com>
wrote:

> Depends on how much we care about performance of mounting multiple volumes
> in parallel.
>
> —
> You are receiving this because your review was requested.
> Reply to this email directly, view it on GitHub
> <https://github.com/kubernetes/kubernetes/pull/66442#issuecomment-408562437>,
> or mute the thread
> <https://github.com/notifications/unsubscribe-auth/AFVgVLRlVxNeP3tL8Ki4TtuSwLVFUnJAks5uK6GGgaJpZM4VXzhH>
> .
>