@cofyc 2018-07-30
Alternative to use reference count to free locks, we can use fixed-size locks, each volume get its lock by `locks[hash(globalPath) % len(locks)]`, then we don't need to free locks. The disadvantage is that different volumes have chance to wait on same lock.

Another way is to use file lock instead of current in-process per string lock, but probably it is not a good idea.