@thockin 2018-07-30
Do we have any understanding of the impact of mutexing all mounts?  With a
single lock, have we run experiments to see the impact?  Can we?  E.g. make
1024 loopback-fs files, time how long it takes to mount and unmount them
all serially single-threaded.  Then shard {2, 4, 8, 16, 32} ways and time
that.  Graph.  That sounds like a day's work and very interesting to learn.

The idea of a fixed set of locks is interesting.  The lifecycle is much
easier to comprehend than refcounts, and honestly even having 4 or 8 locks
is probably sufficient (could be a config if needed, but don't start
there).  Maybe start with number of cores.

Michelle, what do you think?

On Mon, Jul 30, 2018 at 12:17 AM Yecheng Fu <notifications@github.com>
wrote:

> Alternative to use reference count to free locks, we can use fixed-size
> locks, each volume get its lock by locks[hash(globalPath) % len(locks)],
> then we don't need to free locks. The disadvantage is that different
> volumes have chance to wait on same lock.
>
> Another way is to use file lock instead of current in-process per string
> lock, but probably it is not a good idea.
>
> —
> You are receiving this because your review was requested.
> Reply to this email directly, view it on GitHub
> <https://github.com/kubernetes/kubernetes/pull/66442#issuecomment-408769575>,
> or mute the thread
> <https://github.com/notifications/unsubscribe-auth/AFVgVOOECRd299xDpPHLS_grclmaPtOgks5uLrL_gaJpZM4VXzhH>
> .
>