@msau42 2018-07-30
I think having a fixed (or configurable) number of locks to shard is a simpler solution if we determine that these plugins still need these locks.  Ideally, I would like to remove the need for needing to lock mounts in the first place.

* Local volume: we can get rid of the lock once we move to the local attacher interface
* Other volume plugins: I think they use the lock to count mount references to the same volume.  If they implement `MountDevice`, then I think the mount reference counting will no longer be needed because `MountDevice` is serialized on the same volume.  This requires them to implement the Attacher interface, which the local attacher work may benefit them too.  We should sync up with the owners of each relevant volume plugin to see if they can get rid of the mount refcounting.