@msau42 2018-07-30
Looking at the other volume plugins more deeply:
* Only cinder and scaleio plugins are using keymutex for locking across mounts, so those are potentially good candidates for moving to the local attacher interface.  cc @jsafrane for cinder, I don't know who's maintaining scaleio anymore
* iscsi doesn't appear to be using it anymore and the code just needs to be cleaned up
* azure_dd and vsphere seem to be using it for synchronizing Attach across the same node. This needs investigation by respective owners if they can use a sharded locking mechanism.  cc @andyzhangx @abrarshivani